@@ -21,22 +21,22 @@
     [Parameter] public string? PlayTimeSeconds { get; set; }
 
     private string PlayerAvatarUrl => $"https://cravatar.eu/helmavatar/{PlayerName}/64.png";
-    
+
     private string FormattedPlayTime()
     {
         var timeSpan = TimeSpan.FromSeconds(int.Parse(PlayTimeSeconds ?? string.Empty));
         var result = string.Empty;
-        
+
         if (timeSpan.Days > 0)
         {
             result += timeSpan.Days > 1 ? $"{timeSpan.Days} Days " : $"{timeSpan.Days} Day ";
         }
-        
+
         if (timeSpan.Hours > 0)
         {
             result += timeSpan.Hours > 1 ? $"{timeSpan.Hours} Hours " : $"{timeSpan.Hours} Hour ";
         }
-        
+
         if (timeSpan.Minutes > 0)
         {
             result += timeSpan.Minutes > 1 ? $"{timeSpan.Minutes} Minutes " : $"{timeSpan.Minutes} Minute ";
@@ -44,4 +44,5 @@
 
         return result;
     }
+
 }
