@@ -5,12 +5,12 @@
 <MudDialogProvider/>
 <MudSnackbarProvider/>
 
-<MudLayout >
+<MudLayout>
     <MudMainContent>
         <MudContainer MaxWidth="MaxWidth.Large">
-            
-                @Body
-            
+
+            @Body
+
         </MudContainer>
     </MudMainContent>
 </MudLayout>
