@@ -5,6 +5,6 @@ namespace Client.Models;
 public class PlayerQuery
 {
     public List<PlayerSession> PlayerSessions { get; set; } = [];
-    
+
     public int TotalPlaytime { get; set; }
 }
