@@ -8,12 +8,12 @@ public interface IDataService
     public Task<List<Player>> GetOnlinePlayers();
 
     public Task<List<PlayerPlaytime>> GetDayTopPlayers(int limit);
-    
+
     public Task<List<PlayerPlaytime>> GetWeekTopPlayers(int limit);
-    
+
     public Task<List<PlayerPlaytime>> GetMonthTopPlayers(int limit);
-    
+
     public Task<List<PlayerPlaytime>> GetMapTopPlayers(int limit);
-    
+
     public Task<PlayerQuery?> GetPlayerSessions(string playerName);
 }
